feat: complete monorepo interface reorganization for 2025 best practices
- Move package-specific types from shared-types to local packages
- Relocate electron-specific browser types (TabState, CDPMetadata, FavIcon) to apps/electron-app/src/types/browser.ts
- Remove unused interfaces (MemoryNote, MCPToolResult, ContentChunk, etc.) from shared-types
- Eliminate duplicate ExtractionConfig interface in tab-extraction-core
- Consolidate duplicate MCPTool interface in mcp-rag test files
- Clean up empty interface folders from previous reorganization
- Fix all import paths to use correct local vs shared types
- Ensure perfect separation: local types in packages, shared types only for cross-package communication

Results:
- 100% build/lint/typecheck success across all 6 packages
- Zero anti-patterns (no re-exports, no interface pollution)
- Clean architecture following 2025 monorepo best practices
- 83 interfaces perfectly organized across 136 TypeScript files
- Net code reduction: 91 lines removed for cleaner codebase
