+++ apps/electron-app/src/types/ipc-interfaces.ts
@@ -4,10 +4,7 @@
  * Local to electron app - these types are only used within this app
  */
 
-import type {
-  ChatPanelState,
-  ChatPanelRecoveryOptions,
-} from "@vibe/shared-types";
+import type { ChatPanelState, ChatPanelRecoveryOptions } from "./browser";
 
 // App API - System-level operations
 export interface VibeAppAPI {
