+++ packages/shared-types/src/agent/index.ts
@@ -66,47 +66,3 @@ export interface IAgentService extends IAgentProvider {
   // Health monitoring
   performHealthCheck?(): Promise<boolean>;
 }
-
-export interface MemoryNote {
-  id: string;
-  url: string;
-  title: string;
-  synopsis: string;
-  tags: string[];
-  sourceId: string;
-  domain?: string; // Domain extracted from URL
-  createdAt: string;
-  score?: number;
-}
-
-// MCP Tool Result Types
-export interface MCPToolResult {
-  result?:
-    | {
-        content?: Array<{
-          text: string;
-        }>;
-      }
-    | string;
-}
-
-export interface MCPGenerateTextResult {
-  toolResults?: MCPToolResult[];
-}
-
-import type { ContentChunk } from "../content";
-
-export interface MCPSearchMemoryData {
-  type: "memory_discovery" | "content_search" | "recent_memories";
-  query?: string;
-  memories?: MemoryNote[];
-  content_chunks?: ContentChunk[];
-  discovered_domains?: string[];
-}
-
-export interface MCPContentData {
-  type: "content_search";
-  query: string;
-  source_filter?: string;
-  content_chunks: ContentChunk[];
-}
