+++ packages/shared-types/src/content/index.ts
@@ -1,18 +1,4 @@
 /**
- * Content types shared between agent and MCP systems
+ * Content-related shared types
+ * Currently empty - all content types moved to local packages
  */
-
-export interface ContentChunk {
-  id: string;
-  url: string;
-  title?: string;
-  content: string;
-  text?: string; // Alternative content field
-  source_id?: string;
-  similarity?: number;
-  metadata: {
-    title: string;
-    sourceId: string;
-    similarity?: number;
-  };
-}
